Add Navbar documentation links to main layouts
Added a new link for "Navbar" documentation in both the
`DemosMainLayout` and `DocsMainLayout` classes. The new
links are created using the `Link` class with the
appropriate `Href` and `Text` properties set to point to
the respective documentation routes for the Navbar feature.

NOTE: This commit message is auto-generated using GitHub Copilot.